+++ misskey-api/src/streaming/broadcast.rs
@@ -1,1 +1,9 @@
 pub mod emoji_added;
+
+#[cfg(feature = "13-2-3")]
+#[cfg_attr(docsrs, doc(cfg(feature = "13-2-3")))]
+pub mod emoji_updated;
+
+#[cfg(feature = "13-2-3")]
+#[cfg_attr(docsrs, doc(cfg(feature = "13-2-3")))]
+pub mod emoji_deleted;